Added auth.py
