+++ portal/templates/layouts/navbar.html
@@ -2,9 +2,9 @@
 <nav>
   <ul>
     <span>Hello, {{ g.user[2] }}!</span>
-    <li><a href="{{ url_for('course_page') }}">Courses</a></li>
-    <li><a href="{{ url_for('course_page') }}">Dashboard</a></li>
-    <li><a href="{{ url_for('log_out') }}">Log Out</a></li>
+    <li><a href="{{ url_for('courses.course_page') }}">Courses</a></li>
+    <li><a href="{{ url_for('dash') }}">Dashboard</a></li>
+    <li><a href="{{ url_for('auth.log_out') }}">Log Out</a></li>
   </ul>
 </nav>
 {% endif %}
